requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
